Anzeigefehler fr Items mit pic > 15 behoben.

+++ src/de/_13ducks/spacebatz/client/graphics/RenderUtils.java
@@ -47,7 +47,7 @@ public class RenderUtils {
     }
 
     public static float getSourceYForTile(Texture texture, int tile, int tileSize) {
-        return tileSize * (tile / (texture.getImageHeight() / tileSize)) / texture.getImageHeight();
+        return (float) tileSize * (tile / (texture.getImageHeight() / tileSize)) / (float) texture.getImageHeight();
     }
 
     public static float getSourceWidthForTile(Texture texture, int tile, int tileSize) {
